feat: Introduce TripJob, refactor ParcelJob, add type safety
This commit introduces a `TripJob` for background processing related
to trips, refactors the `ParcelJob` to properly destroy the cron job
and includes type safety improvements. Additionally, it defines a new
`TSocketRoutes` type and configures a cascade delete on the
`ParcelHelper` model. It also includes trip data in the socket
emission for parcel requests.